Refactor namespaces for CacheTag and related classes to improve package structure alignment
