+++ config/services.php
@@ -7,8 +7,8 @@ use Neusta\Pimcore\HttpCacheBundle\CacheActivator;
 use Neusta\Pimcore\HttpCacheBundle\CacheInvalidator;
 use Neusta\Pimcore\HttpCacheBundle\CacheInvalidator\OnlyWhenActiveCacheInvalidator;
 use Neusta\Pimcore\HttpCacheBundle\CacheInvalidator\RemoveDisabledTagsCacheInvalidator;
-use Neusta\Pimcore\HttpCacheBundle\CacheTagChecker\ElementCacheTagChecker;
-use Neusta\Pimcore\HttpCacheBundle\CacheTagChecker\StaticCacheTagChecker;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheTagChecker\ElementCacheTagChecker;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheTagChecker\StaticCacheTagChecker;
 use Neusta\Pimcore\HttpCacheBundle\Element\ElementRepository;
 use Neusta\Pimcore\HttpCacheBundle\Element\EventListener\InvalidateElementListener;
 use Neusta\Pimcore\HttpCacheBundle\Element\EventListener\TagElementListener;
