+++ src/CacheTag/CacheTagChecker.php
@@ -1,6 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle;
+namespace Neusta\Pimcore\HttpCacheBundle\CacheTag;
+
+use Neusta\Pimcore\HttpCacheBundle\CacheTag;
 
 interface CacheTagChecker
 {
