+++ src/CacheTag/CacheType.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle;
+namespace Neusta\Pimcore\HttpCacheBundle\CacheTag;
 
 interface CacheType
 {
