+++ src/CacheTag/CacheTypeFactory.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle;
+namespace Neusta\Pimcore\HttpCacheBundle\CacheTag;
 
-use Neusta\Pimcore\HttpCacheBundle\CacheType\CustomCacheType;
-use Neusta\Pimcore\HttpCacheBundle\CacheType\ElementCacheType;
-use Neusta\Pimcore\HttpCacheBundle\CacheType\EmptyCacheType;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheType\CustomCacheType;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheType\ElementCacheType;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheType\EmptyCacheType;
 use Neusta\Pimcore\HttpCacheBundle\Element\ElementType;
 use Pimcore\Model\Element\ElementInterface;
