+++ tests/Unit/Cache/ResponseTagger/RemoveDisabledTagsResponseTaggerTest.php
@@ -3,7 +3,7 @@
 namespace Neusta\Pimcore\HttpCacheBundle\Tests\Unit\Cache\ResponseTagger;
 
 use Neusta\Pimcore\HttpCacheBundle\CacheTag;
-use Neusta\Pimcore\HttpCacheBundle\CacheTagChecker;
+use Neusta\Pimcore\HttpCacheBundle\CacheTag\CacheTagChecker;
 use Neusta\Pimcore\HttpCacheBundle\CacheTags;
 use Neusta\Pimcore\HttpCacheBundle\ResponseTagger;
 use Neusta\Pimcore\HttpCacheBundle\ResponseTagger\RemoveDisabledTagsResponseTagger;
